Use output of std.io.Writer() in Canvas.toPPM()
Previously, Canvas.toPPM() accepted an std.ArrayList to append
characters to a string in PPM format. Now, Canvas.toPPM() directly
accepts the output of std.io.Writer(), the function in the Zig standard
library used to generate a generic writeable context.

Put simply, these changes cut the need for the string previously in the
middle. A PPM-formatted string remains useless unless written to disk as
a PPM file anyway.