Fix `g` command processing in OBJ file parser
Previously, the parser appended faces assigned to a named group in the
Wavefront OBJ file to both that group and the list of all faces. This is
not correct behavior. Now, the parser appends all faces to either their
corresponding named group or the general list of faces that do not
belong to a specific group.