Bound groups
Reduce the number of intersection tests necessary to render a scene by
bounding groups of shapes with a box. When testing ray intersection with
a group of shapes, first test the group's bounding box, and then compute
ray-shape intersection for all shapes within the group if and only if
the ray intersects the bounding box.